fix: more efficient access changes
