+++ packages/server/graphql/public/mutations/createPage.ts
@@ -15,8 +15,14 @@ const createPage: MutationResolvers['createPage'] = async (
     throw new GraphQLError('Can only provider either parentPageId OR teamId')
   }
   const viewerId = getUserId(authToken)
-  const viewer = await dataLoader.get('users').loadNonNull(viewerId)
   const dbParentPageId = parentPageId ? CipherId.fromClient(parentPageId)[0] : undefined
+  const [viewer, parentPage] = await Promise.all([
+    dataLoader.get('users').loadNonNull(viewerId),
+    dbParentPageId ? dataLoader.get('pages').load(dbParentPageId) : null
+  ])
+  if (dbParentPageId && !parentPage) {
+    throw new GraphQLError('Invalid parentPageId')
+  }
 
   const pg = getKysely()
   const topPage = await pg
@@ -32,17 +38,18 @@ const createPage: MutationResolvers['createPage'] = async (
     .limit(1)
     .executeTakeFirst()
   const sortOrder = positionBefore(topPage?.sortOrder ?? ' ')
-
   const page = await getKysely()
     .insertInto('Page')
     .values({
       userId: viewerId,
       parentPageId: dbParentPageId,
+      ancestorIds: parentPage?.ancestorIds.concat(dbParentPageId!) ?? [],
       teamId,
       sortOrder
     })
     .returningAll()
     .executeTakeFirstOrThrow()
+
   analytics.pageCreated(viewer, page.id)
   return {page}
 }
